assistant/ai-assistant-api-old-python/requirements.txt
openai
fastapi
black
tiktoken
isort
websockets
pydantic
langchain
uvicorn
jinja2
faiss-cpu
bs4
unstructured
libmagic
weaviate-client
psycopg2
pyfarmhash

